更新shader（交互shader），更新UI（UI动画，UI黑幕层级优化），更新中文字体，更新受伤效果，场景交互物品的material需要调整（设置成itemMaterial）

+++ Assets/Scripts/FSM/Base/FSMBase.cs
@@ -12,7 +12,7 @@ public abstract class FSMBase : MonoBehaviour
     [Tooltip("追击速度")]
     public float chaseSpeed;
     [Tooltip("血量")]
-    public float HP;
+    public int HP;
     [Tooltip("默认状态编号")]
     public FSMStateID DefaultStateID;
     //只要一个圆形半径就好了
